Solver Servlet now taking input and dishing out found solutions.
This may require a few changes to the project setup. The 'assets' folder
needs to be marked as a source folder (right click -> Build Path -> Use
as Source Folder). The assets folder then needs to be copied to
"WebContent/WEB-INF/", or as I did, it can be symlinked from where it is
now. This isn't an ideal setup, but for now it means /cryptic/solver can
have a "clue" and a "pattern" POSTed to it and it returns any found
solutions.